+++ R/Zdeprecated.R
@@ -1,5 +1,33 @@
 
+## spatstat conflicts
+
+if (!isGeneric("convexhull")) {setGeneric("convexhull", function(x, ...) standardGeneric("convexhull"))}
+setMethod("convexhull", signature(x="SpatVector"), 
+	function(x) {
+		error("convexhull", "terra::convexhull has been removed. Use 'convhull' instead")
+	}
+)
+
+
+if (!isGeneric("perimeter")) {setGeneric("perimeter", function(x, ...) standardGeneric("perimeter"))}
+setMethod("perimeter", signature(x="SpatVector"), 
+	function(x) {
+		error("perimeter", "terra::perimeter has been removed. Use 'perim' instead")
+	}
+)
+
+if (!isGeneric("tiles")) {setGeneric("tiles", function(x,...) standardGeneric("tiles"))}
+setMethod("tiles", signature(x="SpatRaster"), 
+	function(x) {
+		error("tiles", "terra::tiles has been removed. Use 'makeTiles' instead")
+	}
+)
+
+
+
+## tidyverse conflicts
 #not exported
+if (!isGeneric("expand")) {setGeneric("expand", function(x, y, ...) standardGeneric("expand"))}
 setMethod("expand", signature(x="ANY"), 
 	function(...) {
 		warn("expand", "terra::expand has been removed. Use 'extend' instead")
@@ -8,6 +36,7 @@ setMethod("expand", signature(x="ANY"),
 )
 
 #not exported
+if (!isGeneric("near")) {setGeneric("near", function(x, ...) standardGeneric("near"))}
 setMethod("near", signature(x="ANY"), 
 	function(x, ...) {
 		error("near", "near is deprecated. Use 'nearby' instead")
@@ -15,6 +44,7 @@ setMethod("near", signature(x="ANY"),
 )
 
 #not exported
+if (!isGeneric("separate")) {setGeneric("separate", function(x, ...) standardGeneric("separate"))}
 setMethod("separate", signature(x="ANY"), 
 	function(x, ...) {
 		error("separate", "deprecated function. Use 'segregate'")
@@ -24,6 +54,7 @@ setMethod("separate", signature(x="ANY"),
 
 
 #not exported
+if (!isGeneric("pack")) {setGeneric("pack", function(x, ...) standardGeneric("pack"))}
 setMethod("pack", signature(x="ANY"), 
 	function(x, ...) {
 		error("pack", "deprecated function, use 'wrap' instead")
@@ -31,7 +62,7 @@ setMethod("pack", signature(x="ANY"),
 	}
 )
 
-
+## replaced for other reasons
 #not exported
 desc <- function(x) {#
 	error("desc", "deprecated function. Use 'describe'")
@@ -42,5 +73,4 @@ desc <- function(x) {#
 gdal_version <- function() {
 	gdal()
 }
-